Commit para jenny

+++ PreinscripcionAuto/src/main/java/co/edu/udistrital/sga/preinscripcion/auto/services/CursosService.java
@@ -7,7 +7,7 @@ import co.edu.udistrital.sga.preinscripcion.auto.domain.AsignaturaRequerida;
 
 public interface CursosService {
 
-	public List<AsignaturaGrupo> obtenerCursosProgramados(Integer anio, Integer periodo, Integer codigoCarrera);
+	public List<AsignaturaGrupo> obtenerCursosProgramados(Integer anio, Integer periodo, Long codigoCarrera);
 	public List<AsignaturaRequerida> obtenerListaMateriasPreinsDemanda(Integer anio, Integer periodo, Long codCarrera, String estado);
 	public List<AsignaturaRequerida> obtenerListaMateriasPreinsPre(Integer anio, Integer periodo, Long codCarrera, String estado);
 }
